@@ -1,6 +1,6 @@
 package com.applikeysolutions.cosmocalendar.adapter;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.ViewGroup;
 
 import com.applikeysolutions.cosmocalendar.adapter.viewholder.MonthHolder;
