@@ -1,6 +1,6 @@
 package com.applikeysolutions.cosmocalendar.adapter.viewholder;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 import android.widget.TextView;
 
