@@ -1,13 +1,11 @@
 package com.applikeysolutions.cosmocalendar.selection;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import com.applikeysolutions.cosmocalendar.model.Day;
 import com.applikeysolutions.cosmocalendar.selection.criteria.BaseCriteria;
 
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collection;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
