@@ -1,7 +1,7 @@
 package com.applikeysolutions.cosmocalendar.utils;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import com.applikeysolutions.cosmocalendar.model.Day;
 
