@@ -1,8 +1,8 @@
 package com.applikeysolutions.cosmocalendar.view;
 
 import android.content.Context;
-import android.support.annotation.Nullable;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.Nullable;
+import androidx.recyclerview.widget.RecyclerView;
 import android.util.AttributeSet;
 import android.view.animation.Interpolator;
 
