@@ -1,6 +1,6 @@
 package com.applikeysolutions.cosmocalendar.sample;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 
 import com.applikeysolutions.cosmocalendar.utils.SelectionType;
